Rectified IDE errors when joining my code with main
In deleteEnquiry method in Attendee class, the many red lines error is likely just a false positive from the IDE. The code will probably compile and run just fine.

This is what I would consider the first draft of my code, because it is still untested with the whole program.